+++ pkg/grpc/service/studyinfo_endpoints.go
@@ -115,7 +115,13 @@ func (s *studyServiceServer) HasParticipantStateWithCondition(ctx context.Contex
 		}
 
 		cond := types.ExpressionArgFromAPI(req.Condition)
-		if cond.IsExpression() {
+		if cond == nil {
+			return &api.ServiceStatus{
+				Version: apiVersion,
+				Status:  api.ServiceStatus_NORMAL,
+				Msg:     "participant found in study",
+			}, nil
+		} else if cond.IsExpression() {
 			evalCtx := studyengine.EvalContext{
 				ParticipantState: pState,
 			}